# textparse: reject duplicate labels