--- a/pkg/textparse/promparse.go
+++ b/pkg/textparse/promparse.go
@@ -380,7 +380,7 @@ func (p *PromParser) parseLVals() error {
 		case tLName:
 			label := p.l.b[p.l.start:p.l.i]
 			for _, l := range labels {
-				if bytes.Compare(l, label) == 0 {
+				if bytes.Equal(l, label) {
 					return errors.Errorf(`duplicate label "%s"`, string(l))
 				}
 			}