@roidelapluie 2020-01-18
I have researched it.

Unless you use a Builder, labels is happy with duplicate labels (and should be, as it can be used in promql).

In this case, we do
```
var res labels
res.Metric(&res)
```
so we don't check for duplicate and ingest that promptly.